Add AIA.download_range and update animation when screen pressed
- AIA takes a photo every few minutes, download a range of photos and then update using latest.
- Change the animation on press to a full 15fps cycle of the stored images

+++ aia.py
@@ -2,12 +2,94 @@ import os
 import datetime
 import requests
 import os.path
-import datetime
+from datetime import datetime, timedelta
 import requests
 import re
+from bs4 import BeautifulSoup
+
+def listFD(url, ext=''):
+    """
+    Generate a list of file URLs with a specified file extension in a given directory URL.
+
+    Args:
+        url (str): The URL of the directory to list files from.
+        ext (str, optional): The file extension to filter by. Defaults to an empty string (no filtering).
+
+    Returns:
+        list: A list of file URLs in the specified directory with the given file extension.
+
+    Example:
+        file_urls = listFD('https://example.com/files', ext='.txt')
+    """
+    page = requests.get(url).text
+    soup = BeautifulSoup(page, 'html.parser')
+    return [url + '/' + node.get('href') for node in soup.find_all('a') if node.get('href').endswith(ext)]
+
+def daterange(start_date, end_date):
+    for n in range(int((end_date - start_date).days) + 1):
+        yield start_date + timedelta(n)
 
 class AIA:
-    def download_latest_time(self, channel):
+
+    def download_range(start_date, end_date, channel, resolution, download_path):
+        """
+        Download a range of SDO (Solar Dynamics Observatory) images for a specific channel and resolution.
+
+        Args:
+            start_date (datetime): The start date for the range.
+            end_date (datetime): The end date for the range.
+            channel (str): The SDO channel identifier. For example, '0171', '0193', '0304'.
+            resolution (str): The resolution of the images. For example, '512', '1024'.
+            download_path (str): The local directory path where the images will be downloaded.
+
+        Returns:
+            None
+
+        Example:
+            start_date = datetime(2023, 12, 1)
+            end_date = datetime(2023, 12, 17)
+            channel = '0193'
+            resolution = '512'
+            download_path = '/path/to/download'
+            download_range(start_date, end_date, channel, resolution, download_path)
+        """      
+        # https://sdo.gsfc.nasa.gov/assets/img/browse/2023/12/17/20231217_154717_512_0193.jpg 
+        sdo_url = "https://sdo.gsfc.nasa.gov"
+        browse_path = "/assets/img/browse/"
+        
+        for single_date in daterange(start_date, end_date):
+            print(f"Downloading:... {single_date.strftime('%Y-%m-%d')}")
+            dayurl = f"https://sdo.gsfc.nasa.gov/assets/img/browse/{single_date.strftime('%Y/%m/%d')}"
+            files = listFD(dayurl, "jpg")
+            
+            needle = f"{resolution}_{channel}.jpg"
+            filesOfInterest = filter(lambda f: needle in f, files)
+            
+            for file in filesOfInterest:
+                # Determine date "20231217_170741_512_0193.jpg"
+                f = os.path.splitext(os.path.basename(file))[0]
+                split = f.split('_')
+                stamp = split[0]+split[1]
+                # 20231217153753
+                date = datetime.strptime(stamp, '%Y%m%d%H%M%S')
+                if(date < start_date):
+                    continue
+                # Download
+                filename = os.path.join(download_path, f"{os.path.basename(file)}")
+                
+                if(not os.path.isfile(f"{filename}")):
+                    print(f"Download: {file} => {filename}")
+                    try:
+                        response = requests.get(file)
+                        response.raise_for_status()  # Raise an HTTPError for bad responses
+                        with open(filename, 'wb') as file:
+                            file.write(response.content)
+                    except requests.exceptions.RequestException as e:
+                        print(f"Error {file} => {filename} ({e})")
+                else:
+                    print(f"Exists: {file} => {filename}")
+        
+    def download_latest_time(channel):
         """
         Download the latest timestamp from the Solar Dynamics Observatory (SDO) for a specified channel.
 
@@ -16,7 +98,7 @@ class AIA:
                         '1600', '1700', 'HMIB', 'HMII', 'HMID', 'HMIBC', 'HMIIF', 'HMIIC'.
 
         Returns:
-            datetime.datetime: The timestamp of the latest image capture.
+            datetime: The timestamp of the latest image capture.
 
         Notes:
             The function fetches the timestamp data from the 'times' file associated with the specified SDO channel.
@@ -43,13 +125,13 @@ class AIA:
             timestring = match.group(1)
         else:
             # use current time is times file download failed
-            timestring = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
+            timestring = datetime.now().strftime("%Y%m%d_%H%M%S")
 
         # Convert timestamp string to a datetime object
-        timestamp = datetime.datetime.strptime(timestring, "%Y%m%d_%H%M%S")
+        timestamp = datetime.strptime(timestring, "%Y%m%d_%H%M%S")
         return timestamp
 
-    def download_latest_image(self, channel, resolution, download_path):
+    def download_latest_image(channel, resolution, download_path):
         """
         Downloads the latest solar image from the Solar Dynamics Observatory (SDO) for a specified channel and resolution.
         Images are updated per 15 minutes.
@@ -74,23 +156,26 @@ class AIA:
         imageurl = f"{sdo_url}/{latest_path}{latest_file}"
         timesurl = f"{sdo_url}/{latest_path}{latest_times}"
 
-        timestamp = self.download_latest_time(channel)
+        timestamp = AIA.download_latest_time(channel)
         timestring = timestamp.strftime("%Y%m%d_%H%M%S")
 
         # Download the image and save it locally
-        filename = os.path.join(download_path, f"{timestring}-{os.path.basename(imageurl)}")
-        try:
-            response = requests.get(imageurl)
-            response.raise_for_status()  # Raise an HTTPError for bad responses
-            with open(filename, 'wb') as file:
-                file.write(response.content)
-            print(f"Downloaded successfully to: {filename}")
-        except requests.exceptions.RequestException as e:
-            print(f"Error downloading file: {e}")
-
+        filename = os.path.join(download_path, f"{timestring}_{os.path.basename(imageurl)}")
+        filename = filename.replace('latest_','')
+        if(not os.path.isfile(f"{filename}")):
+            print(f"Download: {file} => {filename}")
+            try:
+                response = requests.get(imageurl)
+                response.raise_for_status()  # Raise an HTTPError for bad responses
+                with open(filename, 'wb') as file:
+                    file.write(response.content)
+            except requests.exceptions.RequestException as e:
+                print(f"Error {file} => {filename} ({e})")
+        else:
+            print(f"Exists: {file} => {filename}")
         return filename, timestamp
 
-    def download_latest_video(self, channel, resolution, download_path):
+    def download_latest_video(channel, resolution, download_path):
         """
         Downloads the latest solar video from the Solar Dynamics Observatory (SDO) for a specified channel and resolution.
 
@@ -114,8 +199,8 @@ class AIA:
         imageurl = f"{sdo_url}/{latest_path}{latest_file}"
 
         # use current time is times file download failed
-        timestamp = datetime.datetime.now()
-        timestring = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
+        timestamp = datetime.now()
+        timestring = datetime.now().strftime("%Y%m%d_%H%M%S")
 
         # Download the image and save it locally
         filename = os.path.join(download_path, f"{timestring}-{os.path.basename(imageurl)}")
@@ -136,7 +221,8 @@ if __name__ == "__main__":
     resolution = "512"
     download_path = "./data/"
 
-    print(AIA.download_latest_time(channel))
+    #print(AIA.download_latest_time(channel))
     #download_images(start_date, end_date, channel, resolution, download_path)
     #download_latest_image(channel,resolution,download_path)
     #download_latest_video(channel,resolution,download_path)
+    AIA.download_range(datetime.now()-timedelta(hours=2), datetime.now(), channel, resolution, download_path)
